requirement.txt
fastapi
uvicorn
python-dotenv
openai
reportlab
